Refactor utility files
