--- src/js/util/RouteURL.ts
@@ -1,9 +1,0 @@
-enum RoutesUrls  {
-    HOME = "/",
-    DISCUSSION = "/discussion",
-    BOOKSEARCH ="/book",
-    PROFILE = "/profile",
-    SIGN_IN_LOGIN = "/auth",
-  }
-  
-  export default RoutesUrls;
